+++ halatuju_api/halatuju/settings/base.py
@@ -92,7 +92,7 @@ REST_FRAMEWORK = {
         'rest_framework.parsers.JSONParser',
     ],
     'DEFAULT_PERMISSION_CLASSES': [
-        'rest_framework.permissions.AllowAny',
+        'halatuju.middleware.supabase_auth.SupabaseIsAuthenticated',
     ],
 }